Implement theme hooks.
As discussed in #6, we need a way to load a theme's colors without loading the
entire theme structure, and a way to restrict a theme to a specific portion of
the page.

This commit adds theme hooks, which are CSS classes of the form `theme-*`, to
hold a specific piece of style, such as a `color` or `background-image`. The
layouts and views can use these classes to style particular elements.

This commit also adds app namespaces, which are CSS classes of the form
`app-*`, to isolate a theme to a specific subtree of the page.