Lots of changes.
This is bad form, but we're gonna commit a bunch of changes at once. We're now
using Persona for log in. We're not using Disqus anymore. We have a toolbar. We
can produce content dependent on the user's role. We can add articles to blogs.
We can track comments on blogs.