Rework comment and article partials.
We've introduced a new comment partial, `comments_area`, which includes the form
for creating a new comment. It assumes that the `@comment` instance variable has
been set.

We also reworked the `article` partial to include the `article_summary`
partial. Now passing `summary: true` as a local will produce the output of
`article_summary`.